+++ pages/api/v1/auth/actions/followUser.ts
@@ -8,38 +8,13 @@ import storeApiKeyUsage from '~src/api-middlewares/storeApiKeyUsage';
 import withErrorHandling from '~src/api-middlewares/withErrorHandling';
 import { isValidNetwork } from '~src/api-utils';
 import { followsCollRef } from '~src/api-utils/firestore_refs';
+import updateFollowCounts from '~src/api-utils/updateUserFollowCounts';
 import authServiceInstance from '~src/auth/auth';
 import { MessageType } from '~src/auth/types';
 import getTokenFromReq from '~src/auth/utils/getTokenFromReq';
 import messages from '~src/auth/utils/messages';
 import { firestore_db } from '~src/services/firebaseInit';
 import { IFollowEntry } from '~src/types';
-
-async function updateFollowCounts(userId: number, targetUserId: number, network: string) {
-	const userRef = firestore_db.collection('users').doc(String(userId));
-	const targetUserRef = firestore_db.collection('users').doc(String(targetUserId));
-	const userDoc = await userRef.get();
-	const targetUserDoc = await targetUserRef.get();
-	if (!userDoc.exists || !targetUserDoc.exists) {
-		throw new Error('User document not found');
-	}
-
-	const batch = firestore_db.batch();
-
-	batch.update(userRef, {
-		[`followings_count.${network}`]: Math.max((userDoc.data()?.followings_count?.[network] || 0) + 1, 0)
-	});
-	batch.update(targetUserRef, {
-		[`followers_count.${network}`]: Math.max((targetUserDoc.data()?.followers_count?.[network] || 0) + 1, 0)
-	});
-
-	try {
-		await batch.commit();
-	} catch (error) {
-		console.error('Error committing batch:', error);
-		throw new Error('Failed to update follow counts');
-	}
-}
 
 async function handler(req: NextApiRequest, res: NextApiResponse<MessageType>) {
 	storeApiKeyUsage(req);
@@ -81,7 +56,7 @@ async function handler(req: NextApiRequest, res: NextApiResponse<MessageType>) {
 				updated_at: new Date()
 			});
 			res.status(200).json({ message: 'User followed' });
-			await updateFollowCounts(user.id, userIdToFollow, network);
+			await updateFollowCounts({ isFollowing: true, network, targetUserId: userIdToFollow, userId: user.id });
 			return;
 		}
 	} catch (error) {
@@ -107,7 +82,7 @@ async function handler(req: NextApiRequest, res: NextApiResponse<MessageType>) {
 	//TODO: send notification to the user followed
 
 	res.status(200).json({ message: 'User followed' });
-	await updateFollowCounts(user.id, userIdToFollow, network);
+	await updateFollowCounts({ isFollowing: true, network, targetUserId: userIdToFollow, userId: user.id });
 	return;
 }
